+++ custom/champ/tests/__init__.py
@@ -4,15 +4,12 @@ import postgres_copy
 import sqlalchemy
 import os
 
-from django.test.testcases import TestCase
-from django.test.client import RequestFactory
 from django.test.utils import override_settings
 
 from corehq.apps.domain.models import Domain
 from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.userreports.models import StaticDataSourceConfiguration
 from corehq.apps.userreports.util import get_indicator_adapter, get_table_name
-from corehq.apps.users.models import WebUser
 from corehq.sql_db.connections import connection_manager, UCR_ENGINE_ID
 
 
@@ -22,8 +19,8 @@ def setUpModule():
     )
     _call_center_domain_mock.start()
 
+    domain = create_domain('champ-cameroon')
     with override_settings(SERVER_ENVIRONMENT='production'):
-        domain = create_domain('champ-cameroon')
 
         configs = StaticDataSourceConfiguration.by_domain(domain.name)
         adapters = [get_indicator_adapter(config) for config in configs]
@@ -40,24 +37,13 @@ def setUpModule():
                 table_name = get_table_name(domain.name, file_name[:-4])
                 table = metadata.tables[table_name]
                 postgres_copy.copy_from(f, table, engine, format='csv', null='', header=True)
-
     _call_center_domain_mock.stop()
 
 
-class ChampTestCase(TestCase):
-
-    def setUp(self):
-        self.run_july_third_test = True
-        self.factory = RequestFactory()
-        domain = Domain.get_or_create_with_name('champ-cameroon')
-        domain.is_active = True
-        domain.save()
-        self.domain = domain
-        user = WebUser.all().first()
-        if not user:
-            user = WebUser.create(domain.name, 'test', 'passwordtest')
-        user.is_authenticated = True
-        user.is_superuser = True
-        user.is_authenticated = True
-        user.is_active = True
-        self.user = user
+def tearDownModule():
+    _call_center_domain_mock = mock.patch(
+        'corehq.apps.callcenter.data_source.call_center_data_source_configuration_provider'
+    )
+    _call_center_domain_mock.start()
+    Domain.get_by_name('champ-cameroon').delete()
+    _call_center_domain_mock.stop()
