+++ custom/champ/tests/test_pva_chart.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 import json
 import mock
-from custom.champ.tests import ChampTestCase
+from custom.champ.tests.utils import ChampTestCase
 from custom.champ.views import PrevisionVsAchievementsView
 
 from django.core.urlresolvers import reverse
