+++ custom/champ/tests/test_service_uptake.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
-from custom.champ.tests import ChampTestCase
+from custom.champ.tests.utils import ChampTestCase
 from custom.champ.views import ServiceUptakeView
 
 import json
